# backend/Swagger.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: API Backend de un restaurante de comida típica
  contact:
    email: dmesa@gmail.com
  version: 1.0.0
...
servers:
  - url: https://localhost:3000/
  - url: http://localhost:3000/
...
paths:
...
  /deleteUser:
     delete:
      tags:
        - User
      summary: Eliminar usuario (solo Administrador)
      description: Elimina un usuario por la identificacion.
      parameters:
        - name: id
          in: query
          example: id=1
          description: Necesitamos la identificacion del usuario
          required: true
          schema:
            type: string
      operationId: deleteUser
      responses:
        200:
          description: "Se ha eliminado el usuario ${id} de manera correcta"
        406:
          description: "No se ha podido eliminar el usuario ya que tiene órdenes asociadas"
        400:
          $ref: "#/components/responses/UnauthorizedAdmin"
        default:
          $ref: "#/components/responses/Default"      
  /getUser:
    get:
      tags:
        - User
      summary: Obtiene todos los usuarios registrados en la base de datos
      description: Devuelve una lista de usuarios
      operationId: getUser
      parameters:
        - name: search
          in: query
          description: busca todos los usuarios
          schema:
            type: string
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUser'
        400:
          description: Error al buscar la información
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
  /updateUser/:id:
    put:
      tags:
        - User
      summary: Actualiza un usuario (solo Administrador)
      description: Actualiza toda la información del usuario con el id del mismo
      operationId: updateUser
      parameters:
        - name: id
          in: query
          example: id=1
          description: Necesitamos la identificación del usuario
          required: true
          schema:
            type: string
      requestBody:
        description: updateUser
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateUser'
        required: true
      responses:
        200:
          description: "El usuario identificado con el id: ${id} fue actualizado correctamente"
        400:
          description: La información ingresada está incompleta o mal redactada
        default:
          $ref: "#/components/responses/Default"

#ORDENES
...
components:
  schemas:
...
    getUser:
      type: object
      properties:
        user_id:
          type: integer
          example: 7
        username:
          type: string
          example: daniel
        password:
          type: string
          example: waynerooney
        full_name: 
          type: string
          example: Wayne Rooney Mccallister
        email: 
          type: string
          example: rooney10a@gmail.com
        phone:
          type: integer
          example: 411121
        delivery_address: 
          type: string
          example: Amaga por la loma de abajo
        is_admin: 
          type: integer
          example: 0
        is_disabled: 
          type: integer
          example: 0     
    updateUser:
      type: object
      properties:
        user_id:
          type: integer
          example: 7
        username:
          type: string
          example: jack
        password:
          type: string
          example: astonvilla
        full_name: 
          type: string
          example: Jack Grealish
        email: 
          type: string
          example: jackgrealish23a@gmail.com
        phone:
          type: integer
          example: 568791245
        delivery_address: 
          type: string
          example: Villa Park 45-09      
          
          
#Schema-Products
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "An error has ocurred: <error>"
      required:
        - error
    Send:
      type: string
      example: "Error message"
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    
      

  # Reused Responses
  responses:
...
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
    UnauthorizedAdmin:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
            Not Admin:
              summary: Unauthorized - Not an admin
              value: Unauthorized - Not an admin
...
security:
  - bearerAuth: []          